refactor: 백엔드 코드 리팩토링
