+++ .env.example
@@ -1,9 +1,15 @@
 # 한국은행 API키
-ECOS_API_KEY="YOUR_API_KEY"
+ECOS_API_KEY="YOUR_ECOS_API_KEY"
 
 # MongoDB Connection Settings
-MONGO_URI="Your_Mongo_URI"
-MONGO_DB_NAME="Your_Mongo_DB_Name"
+MONGO_URI="YOUR_MONGO_URI"
+MONGO_DB_NAME="YOUR_MONGO_DB_NAME"
+
+# Redis Connection Settings
+REDIS_URI="YOUR_REDIS_URI"
+
+# Kafka Bootstrap Servers
+KAFKA_BROKER="YOUR_BOOTSTRAP_SERVER"
 
 # FastAPI Logging Level
 DEBUG=1
