+++ backend/domains/users/services/auth_service.py
@@ -6,7 +6,7 @@ from uuid import uuid4
 from fastapi import Request, WebSocket
 from passlib.context import CryptContext
 from domains.users.schemes.user_info import UserInfo
-from modules.database import MongoDB, Redis
+from modules.redis import Redis
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 SESSION_TTL = 3600  # 1 hour
